requirements.txt
keras==2.10.0
matplotlib==3.7.1
numpy==1.23.5
opencv_python==4.8.0.76
pygame==2.5.1
tensorflow_gpu==2.10.0
tensorflow_intel==2.13.0
